Fix the GIC CPU interface address for zynqmp board
The first 4kB of the GIC CPU interface is aliased over the first 64kB
area it is mapped on (0xf9020000 to 0xf902ffff for the zynqmp).
The second 4kB (the C_DIR register) is accessible from 0xf9030000. In
order to have a continuous 8kB area, we use 0xf902f000 as the base
address.

See
  http://lkml.kernel.org/r/1442142873-20213-2-git-send-email-marc.zyngier and
  https://lists.xen.org/archives/html/xen-devel/2015-10/msg01068.html
for more info.

Signed-off-by: Luc MICHEL <luc.michel@git.antfield.fr>
Signed-off-by: Edgar E. Iglesias <edgar.iglesias@xilinx.com>